fix active tab state detection, reorganize files, cleanup data
