+++ components/Footer/PlatformTile.tsx
@@ -2,9 +2,8 @@ import { createElement, useContext, type ComponentType } from 'react';
 import { StyleSheet, View, type ViewStyle } from 'react-native';
 
 import { A, colors, darkColors, P } from '~/common/styleguide';
+import { type IconProps } from '~/components/Icons';
 import CustomAppearanceContext from '~/context/CustomAppearanceContext';
-
-import { type IconProps } from './Icons';
 
 type PlatformProps = {
   name: string;
