+++ components/Head/StructuredData.tsx
@@ -2,7 +2,7 @@ type Props = {
   data: Record<string, any>;
 };
 
-export function StructuredData({ data }: Props) {
+export default function StructuredData({ data }: Props) {
   return (
     <script
       key="structured-data"
