+++ pages/_document.tsx
@@ -1,7 +1,7 @@
 import Document, { Html, Head, Main, NextScript, type DocumentContext } from 'next/document';
 
-import GoogleAnalytics from '~/components/GoogleAnalytics';
-import { StructuredData } from '~/components/StructuredData';
+import GoogleAnalytics from '~/components/Head/GoogleAnalytics';
+import StructuredData from '~/components/Head/StructuredData';
 
 function DirectoryWebsite() {
   return (
